pyproject.toml
[project]
name = "osl-eln-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.6",
    "langchain-anthropic>=1.3.1",
    "langchain-google-genai>=4.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.7",
    "oold>=0.11.4",
    "opensemantic-lab>=0.6.1.post2000",
    "osw",
    "python-dotenv>=1.2.1",
]

[tool.uv.sources]
osw = { git = "https://github.com/OpenSemanticLab/osw-python" }
